+++ kite/src/main/kotlin/com/tang/kite/wrapper/where/AbstractHavingWrapper.kt
@@ -12,13 +12,7 @@ import java.util.function.Consumer
 /**
  * @author Tang
  */
-abstract class AbstractHavingWrapper<R, T>(
-
-    private val wrapper: Wrapper<T>,
-
-    private val conditions: MutableList<LogicalStatement> = mutableListOf()
-
-): AbstractConditionWrapper<R, T>(conditions) {
+abstract class AbstractHavingWrapper<R, T>(private val wrapper: Wrapper<T>, ): AbstractConditionWrapper<R, T>() {
 
     protected var havingInstance: R? = null